requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.41.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
